+++ opencascade-hs/src/OpenCascade/BRepBuilderAPI/Sewing.hs
@@ -6,6 +6,9 @@ module OpenCascade.BRepBuilderAPI.Sewing
 , add
 , perform
 , sewedShape
+, nbFreeEdges
+, nbContigousEdges
+, nbMultipleEdges
 ) where 
 
 import qualified OpenCascade.TopoDS.Types as TopoDS
@@ -13,7 +16,7 @@ import OpenCascade.TopoDS.Internal.Destructors (deleteShape)
 import OpenCascade.BRepBuilderAPI.Internal.Destructors (deleteSewing)
 import OpenCascade.BRepBuilderAPI.Types (Sewing)
 import Foreign.Ptr (Ptr)
-import Foreign.C (CBool (..), CDouble (..))
+import Foreign.C (CBool (..), CDouble (..), CInt (..))
 import OpenCascade.Internal.Bool (boolToCBool)
 import Data.Acquire (Acquire, mkAcquire)
 import Data.Coerce (coerce)
@@ -34,5 +37,19 @@ foreign import capi unsafe "hs_BRepBuilderAPI_Sewing.h hs_BRepBuilderAPI_Sewing_
 sewedShape :: Ptr Sewing -> Acquire (Ptr TopoDS.Shape)
 sewedShape sewing = mkAcquire (rawSewedShape sewing) (deleteShape)
 
+foreign import capi unsafe "hs_BRepBuilderAPI_Sewing.h hs_BRepBuilderAPI_Sewing_nbFreeEdges" rawNbFreeEdges:: Ptr Sewing -> IO (CInt)
+
+nbFreeEdges :: Ptr Sewing -> IO Int
+nbFreeEdges = fmap fromIntegral <$> rawNbFreeEdges
 
 
+foreign import capi unsafe "hs_BRepBuilderAPI_Sewing.h hs_BRepBuilderAPI_Sewing_nbContigousEdges" rawNbContigousEdges:: Ptr Sewing -> IO (CInt)
+
+nbContigousEdges :: Ptr Sewing -> IO Int
+nbContigousEdges = fmap fromIntegral <$> rawNbContigousEdges
+
+
+foreign import capi unsafe "hs_BRepBuilderAPI_Sewing.h hs_BRepBuilderAPI_Sewing_nbMultipleEdges" rawNbMultipleEdges:: Ptr Sewing -> IO (CInt)
+
+nbMultipleEdges :: Ptr Sewing -> IO Int
+nbMultipleEdges = fmap fromIntegral <$> rawNbMultipleEdges